graphrag_demo/rag_evals/requirements.txt
click>=8.0.0
datasets>=2.14.0
huggingface-hub>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
ragas>=0.3.5